fixed GH2 projection bug when the initial matrix is zoomed in.

+++ include/GH1D.h
@@ -36,9 +36,9 @@ public:
   GH1D* GetPrevious() const;
   GH1D* GetNext() const;
 
-  GH1D* Project(int bin_low, int bin_high) const;
-  GH1D* Project_Background(int bin_low, int bin_high,
-                           int bg_bin_low, int bg_bin_high,
+  GH1D* Project(double bin_low, double bin_high) const;
+  GH1D* Project_Background(double bin_low, double bin_high,
+                           double bg_bin_low, double bg_bin_high,
                            kBackgroundSubtraction mode = kRegionBackground) const;
 
 private:
